python/requirements.txt
web3==6.20.1
pandas==2.2.2
matplotlib==3.9.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.7.1

